refactor #66: 전체 삭제 컨펌 모달 컴포넌트로 분리
