+++ fe/src/components/Main/Cart.module.css
@@ -168,30 +168,3 @@
   -webkit-user-select: none;
   user-select: none;
 }
-
-.ModalContent {
-  text-wrap: nowrap;
-  margin-bottom: 20px;
-}
-
-.ButtonContainer {
-  width: 100%;
-  display: flex;
-  gap: 10px;
-}
-
-.ConfirmButton {
-  width: 100%;
-  background-color: #e82444;
-  color: white;
-  font-size: 16px;
-  padding: 10px;
-}
-
-.CancelButton {
-  width: 100%;
-  background-color: black;
-  color: white;
-  font-size: 16px;
-  padding: 10px;
-}
